fixed for PR comments
