+++ Resolution/Resolution/Parser/ChainParser/RecognisedValue.cs
@@ -1,0 +1,10 @@
+﻿namespace Resolution.Parser.ChainParser
+{
+    public enum RecognisedValue
+    {
+        NotRecognised,
+        Keyword,
+        SentenceTyped,
+        EndOfSentence
+    }
+}
