+++ Resolution/Resolution/Parser/Decorators/AbstractTextDecorator.cs
@@ -4,14 +4,18 @@
     {
         private ITextDecorator component;
 
+        public string Text 
+        { 
+            get
+            {
+                return this.Decorate(this.component.Text);
+            }
+        }
+
         public AbstractTextDecorator(ITextDecorator component)
         {
             this.component = component;
         }
         protected abstract string Decorate(string text);
-        public string GetText()
-        {
-            return this.Decorate(this.component.GetText());
-        }
     }
 }
