+++ Resolution/Resolution/Parser/Decorators/BasicText.cs
@@ -4,13 +4,13 @@ namespace Resolution.Parser.Decorators
 {
     public class BasicText : ITextDecorator
     {
-        private string Text { get; set; }
+        private string _text;
+        public string Text { get => this._text; }
         public BasicText(string text)
         {
             if (string.IsNullOrEmpty(text))
                 throw new ArgumentException("Text cannot be null or empty");
-            this.Text = text;
+            this._text = text;
         }
-        public string GetText() => this.Text;
     }
 }
